+++ src/ishowmanager.h
@@ -15,9 +15,9 @@ class Track;
 
 class IShowManager {
 public:
-    void playAtTime(int time);
+    void playAtTime(int64_t time);
     void createTracksForDevices(Array<device::ILogicalDevice *> &devices);
-    virtual void generateTracksActions(unsigned long time) = 0;
+    virtual void generateTracksActions(int64_t time) = 0;
 
     Array<Track*> tracks;
 };
